Add sb regressor docs, rename sb_bagging_classifier to sb_bagging
